+++ ProyectosSDL/HolaSDL/EnlargenAward.cpp
@@ -4,7 +4,7 @@
 
 /// Public
 /// Constructor
-EnlargenAward::EnlargenAward(float32 x, float32 y, float32 width, float32 height, float32 speed, float32 framerate, Texture *texture)
+EnlargenAward::EnlargenAward(float32 x, float32 y, float32 width, float32 height, float32 speed, uint framerate, Texture *texture)
     : Award(x, y, width, height, speed, framerate, texture) {}
 
 /// Public Virtual
